+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Common task variables have not been set, fixed that.
+
 ## [0.10.3] - 2022-01-26
 
 ### Fixed