+++ CHANGELOG.md
@@ -1,9 +1,13 @@
-# Change log for DscResource.DocGenerator
+# Changelog for DscResource.DocGenerator
 
 The format is based on and uses the types of changes according to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Changed
+
+- Updated pipelines files to the latest in Sampler.
 
 ## [0.10.3] - 2022-01-26