+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Updated pipelines files to the latest in Sampler.
+- Fix missing verbose message in function `Invoke-Git`.
+- Fix variable reference for localized strings so they passed HQRM tests.
+- Fix statement so function passed HQRM tests.
 
 ## [0.10.3] - 2022-01-26