+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tasks now uses `Set-SamplerTaskVariable` from the module Sampler to set
   the common build task variables.
 
+### Fixed
+
+- Common task variables have not been set, fixed that.
+
 ## [0.10.3] - 2022-01-26
 
 ### Fixed